+++ R/textstat_sentiment.R
@@ -5,14 +5,14 @@
 #' Compute sentiment scores from tokens or document-feature matrices, based on
 #' assigned categories (types or features) of positive, negative, and neutral
 #' sentiment. Several formulas are available, or the user can supply a new one.
-#' @param x a [dfm] or [tokens] object containing sentiment types or features
-#' @param pos,neg character; names of the dictionary keys or features 
-#'   for positive and negative sentiment, respectively
-#' @param neut character; (optional) name of the neutral sentiment feature or 
-#'   key
+#' @param x a character, [corpus], [tokens], or [dfm] object containing 
+#'   text, tokens, or features whose sentiment will be scored
+#' @param dictionary a [dictionary] that has [polarity] set, indicating which
+#'   keys are associated with positive, negative, and (optionally) neutral 
+#'   sentiment
 #' @param fun function; the formula for computing sentiment, which must refer to
-#'   `pos`, `neg`, and (optionally) `neut`.  Default is the Lowe et al (2001)
-#'   "logit" scale which is the log of (positive / negative) counts.  See
+#'   `pos`, `neg`, and (optionally) `neut`.  The default is the "logit" scale
+#'   (Lowe et al 2011) which is the log of (positive / negative) counts.  See
 #'   [sentiment-functions] for details and for additional available functions,
 #'   as well as details on how to supply custom functions.
 #' @param ... additional arguments passed to `fun`
@@ -23,58 +23,63 @@
 #'   Quarterly_, 36(1), 123–155.
 #'   <http://doi.org/10.1111/j.1939-9162.2010.00006.x>
 #' @examples
-#' toks <- tail(data_corpus_inaugural, n = 5) %>%
-#'     tokens(remove_punct = TRUE, remove_numbers = TRUE) %>%
-#'     tokens_lookup(dictionary = data_dictionary_LSD2015, nomatch = "_other_")
-#' textstat_sentiment(toks)
-#' textstat_sentiment(toks, smooth = 1)
-#' textstat_sentiment(toks, fun = sent_relpropdiff)
-#' 
+#' corp <- tail(data_corpus_inaugural, n = 5)
+#' toks <- tokens(corp)
 #' dfmat <- dfm(toks)
-#' textstat_sentiment(dfmat)
+#' polar1 <- list(pos = "positive", neg = "negative")
+#' polar2 <- list(pos = c("positive", "neg_negative"),
+#'                neg = c("negative", "neg_positive"))
+#'                
+#' polarity(data_dictionary_LSD2015) <- polar1
+#' textstat_sentiment(corp, dictionary = data_dictionary_LSD2015)
+#' textstat_sentiment(toks, dictionary = data_dictionary_LSD2015)
+#' textstat_sentiment(dfmat, dictionary = data_dictionary_LSD2015)
 #' 
-#' # works with multiple types associated with each sentiment category
-#' textstat_sentiment(toks, pos = c("positive", "neg_negative"),
-#'                    neg = c("negative", "neg_positive"))
-#'
+#' polarity(data_dictionary_LSD2015) <- polar2
+#' textstat_sentiment(corp, dictionary = data_dictionary_LSD2015)
+#' textstat_sentiment(toks, dictionary = data_dictionary_LSD2015)
+#' textstat_sentiment(corp, dictionary = data_dictionary_LSD2015)
+#' textstat_sentiment(dfmat, dictionary = data_dictionary_LSD2015)
+#' 
 #' # with a user-supplied function
 #' sent_fn <- function(x) (x[, "pos"] - x[, "neg"]) / rowSums(x) * 100
-#' textstat_sentiment(toks, fun = sent_fn)
-textstat_sentiment <- function(x, pos = "positive", neg = "negative", 
-                               neut = NULL, fun = sent_logit, ...) {
+#' textstat_sentiment(toks, data_dictionary_LSD2015, fun = sent_fn)
+textstat_sentiment <- function(x, dictionary, fun = sent_logit, ...) {
     UseMethod("textstat_sentiment")
 }
 
 #' @export
-textstat_sentiment.default <-  function(x, pos = "positive", neg = "negative", 
-                                        neut = NULL, fun = sent_logit, ...) {
+textstat_sentiment.default <-  function(x, dictionary, fun = sent_logit, ...) {
     stop(friendly_class_undefined_message(class(x), "textstat_sentiment"))
 }
 
 #' @export
-textstat_sentiment.tokens <- function(x, pos = "positive", neg = "negative", 
-                                      neut = NULL, fun = sent_logit, ...) {
-    textstat_sentiment(dfm(x), pos, neg, neut, fun = fun, ...)
+textstat_sentiment.character <- function(x, ...) {
+    textstat_sentiment(corpus(x), ...)
 }
 
 #' @export
-textstat_sentiment.dfm <- function(x, pos = "positive", neg = "negative",
-                                   neut = NULL, fun = sent_logit, ...) {
-    if (!all(c(pos, neg, neut) %in% featnames(x)))
-        stop("pos, neg, and neut must match features in x")
+textstat_sentiment.corpus <- function(x, ...) {
+    textstat_sentiment(tokens(x), ...)
+}
 
-    # get the categories that are not sentiment
-    x_other <- dfm_remove(x, c(pos, neg, neut))
-    colnames(x_other) <- rep("other", nfeat(x_other))
-    x_other <- dfm_compress(x_other, margin = "features")
+#' @export
+textstat_sentiment.tokens <- function(x, dictionary, ...) {
+    dict <- get_polarity_dictionary(dictionary)
+    poldict <- dictionary(polarity(dict))
+    polarity(poldict) <- polarity(dict)
     
-    x <- dfm_lookup(x, dictionary(list(
-        "pos" = pos, "neg" = neg, "neut" = as.character(neut)
-    )))
-    
-    x <- cbind(x, x_other)
-    
-    result <- fun(x, ...)
+    tokens(x) %>%
+        tokens_lookup(dictionary = dict, nomatch = "other") %>%
+        dfm() %>%
+        textstat_sentiment(dictionary = poldict, ...)
+}
+
+#' @export
+textstat_sentiment.dfm <- function(x, dictionary, fun = sent_logit, ...) {
+    dict <- get_polarity_dictionary(dictionary)
+
+    result <- fun(dfm_lookup(x, dict, nomatch = "other"), ...)
     result <- convert(as.dfm(result), to = "data.frame")
     names(result)[2] <- "sentiment"
     
@@ -85,6 +90,112 @@ textstat_sentiment.dfm <- function(x, pos = "positive", neg = "negative",
     result
 }
 
+
+# polarity setting and checking functions --------------
+
+#' Set or get the sentiment polarity of a dictionary
+#'
+#' Set or retrieve the polarity of a [dictionary] object for the purposes of
+#' sentiment analysis.  Polarity consists of a set of dictionary keys that are
+#' associated with positive, negative, and (optionally) neutral categories for
+#' use in [textstat_sentiment()].
+#' 
+#' A dictionary may have only one set of polarities at a time, but may be
+#' changed as needed.
+#' @param x a [dictionary] object
+#' @return `polarity()` returns the polarity as a list.
+#' @keywords dictionary textstat utility
+#' @export
+#'
+#' @examples
+#' simpledict <- dictionary(list(
+#'     happy = c("happy", "jubilant", "exuberant"),
+#'     sad = c("sad", "morose", "down")
+#' ))
+#' polarity(simpledict)
+#' polarity(simpledict) <- list(pos = "happy", neg = "sad")
+#' polarity(simpledict)
+#' 
+#' # can list multiple keys
+#' polarity(data_dictionary_LSD2015) <- list(
+#'     pos = c("positive", "neg_negative"),
+#'     neg = c("negative", "neg_positive")
+#' )
+#' polarity(data_dictionary_LSD2015)
+polarity <- function(x) {
+    UseMethod("polarity")
+}
+
+#' @export
+polarity.dictionary2 <- function(x) {
+    x@meta$object$polarity
+}
+
+#' @rdname polarity
+#' @param value list consisting of named character vectors `pos`, `neg`, and
+#'   (optionally) `neut` corresponding to positive, negative, and neutral 
+#'   sentiment categories respectively.  Each element may contain multiple 
+#'   key names.  The `neut` category is optional but `pos` and `neg` must be
+#'   supplied.
+#' @return `polarity<-` sets the dictionary's polarity.
+#' @export
+"polarity<-" <- function(x, value) {
+    UseMethod("polarity<-")
+}
+
+#' @export
+"polarity<-.dictionary2" <- function(x, value) {
+    if (!setequal(union(c("pos", "neg", "neut"), names(value)), 
+                  c("pos", "neg", "neut")) ||
+        !is.list(value)) {
+        stop("value must be a list of 'pos', 'neg', and (optionally) 'neut'",
+            call. = FALSE)
+    }
+    
+    x@meta$object$polarity <- value
+    x
+}
+
+#' Get a standard polarity dictionary for sentiment analysis
+#' 
+#' Checks and standardizes a [dictionary] object with its [polarity] set, so
+#' that the polarity categories are standardized into the keys `pos`, `neg`, and
+#' (optionally) `neut`.  Also checks that the dictionary contains all of the
+#' keys named in the polarity object.  (It is necessary to check here since the
+#' dictionary could have been subset after creation.)
+#' @param dictionary a \pkg{quanteda} [dictionary]
+#' @return a single-level [dictionary] with keys `pos`, `neg`, and (optionally)
+#'   `neut`.
+#' @keywords internal
+get_polarity_dictionary <- function(dictionary) {
+    poles <- polarity(dictionary)
+    
+    # check the poles
+    if (is.null(poles)) {
+        stop("polarity is not set for this dictionary; see ?polarity", 
+             call. = FALSE)
+    }
+    if (! all(polematch <- unlist(poles, use.names = FALSE) %in%
+              names(dictionary)) ) {
+        stop(polematch[!polematch], " key not found in this dictionary",
+             cqll. = FALSE)
+    }
+    
+    # standardize the dictionary
+    dictlist <- list(
+        pos = unlist(dictionary[poles$pos], use.names = FALSE),
+        neg = unlist(dictionary[poles$neg], use.names = FALSE),
+        neut = unlist(dictionary[poles$neut], use.names = FALSE)
+    )
+    dict <- dictionary(dictlist[!sapply(dictlist, is.null)])
+    
+    # set the polarity to the keys
+    newpoles <- list(pos = "pos", neg = "neg")
+    if (!is.null(dictlist$neut)) newpoles <- c(newpoles, list(neut = "neut"))
+    polarity(dict) <- newpoles
+
+    return(dict)
+}
 
 # sentiment formula functions --------------